+++ src/path_finders.rs
@@ -3,10 +3,14 @@ use std::fs;
 use std::process::Command;
 use walkdir::{DirEntry, WalkDir};
 
-pub fn find_all_paths(root: &str) -> Vec<(String, FileType)> {
+pub fn find_all_paths(root: &str, directories_only: bool) -> Vec<(String, FileType)> {
     let mut result: Vec<(String, FileType)> = Vec::new();
     for entry in WalkDir::new(root).into_iter().filter_map(|e| e.ok()) {
         if let Ok(meta) = entry.metadata() {
+            if directories_only && !meta.is_dir() {
+                continue
+            }
+
             if let Some(path) = entry.path().to_str() {
                 let path = path.to_string();
                 if &path != root {
@@ -26,7 +30,7 @@ fn is_hidden(entry: &DirEntry) -> bool {
         .unwrap_or(false)
 }
 
-pub fn find_non_hidden_paths(root: &str) -> Vec<(String, FileType)> {
+pub fn find_non_hidden_paths(root: &str, directories_only: bool) -> Vec<(String, FileType)> {
     let walker = WalkDir::new(root).into_iter();
     let mut result: Vec<(String, FileType)> = Vec::new();
 
@@ -35,6 +39,9 @@ pub fn find_non_hidden_paths(root: &str) -> Vec<(String, FileType)> {
         .filter_map(|e| e.ok())
     {
         if let Ok(meta) = entry.metadata() {
+            if directories_only && !meta.is_dir() {
+                continue
+            }
             if let Some(path) = entry.path().to_str() {
                 let path = path.to_string();
                 if &path != root {
@@ -46,8 +53,21 @@ pub fn find_non_hidden_paths(root: &str) -> Vec<(String, FileType)> {
     result
 }
 
-pub fn find_non_git_ignored_paths(root: &str) -> Vec<(String, FileType)> {
-    if let Ok(git_output) = Command::new("git").arg("ls-files").arg(root).output() {
+pub fn find_non_git_ignored_paths(root: &str, directories_only: bool) -> Vec<(String, FileType)> {
+    let mut git_command = Command::new("git");
+    if directories_only {
+        git_command
+            .arg("ls-tree")
+            .arg("-r")
+            .arg("-d")
+            .arg("--name-only")
+            .arg("HEAD")
+            .arg(root);
+    } else {
+        git_command.arg("ls-files").arg(root);
+    };
+
+    if let Ok(git_output) = git_command.output() {
         if git_output.status.success() {
             if let Ok(paths_buf) = String::from_utf8(git_output.stdout) {
                 return paths_buf
@@ -63,5 +83,5 @@ pub fn find_non_git_ignored_paths(root: &str) -> Vec<(String, FileType)> {
         }
     }
 
-    return find_non_hidden_paths(root);
+    return find_non_hidden_paths(root, directories_only);
 }